PullRequest: sulong/219

+++ projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/util/ProcessUtil.java
@@ -42,7 +42,7 @@ import com.oracle.truffle.llvm.test.options.TestOptions;
 public class ProcessUtil {
 
     private static final int BUFFER_SIZE = 1024;
-    private static final int PROCESS_WAIT_TIMEOUT = 5000;
+    private static final int PROCESS_WAIT_TIMEOUT = 20000;
 
     /**
      * This class represents the result of a native command executed by the operating system.
@@ -155,9 +155,12 @@ public class ProcessUtil {
         }
         try {
             Process process = Runtime.getRuntime().exec(command);
-            process.waitFor(PROCESS_WAIT_TIMEOUT, TimeUnit.MILLISECONDS);
+            boolean success = process.waitFor(PROCESS_WAIT_TIMEOUT, TimeUnit.MILLISECONDS);
             String readError = readStreamAndClose(process.getErrorStream());
             String inputStream = readStreamAndClose(process.getInputStream());
+            if (!success) {
+                throw new AssertionError("timeout running command: " + command);
+            }
             int llvmResult = process.exitValue();
             process.destroyForcibly();
             return new ProcessResult(command, llvmResult, readError, inputStream);
